requirements.txt
fastapi
uvicorn
python-dotenv
redis
supabase
google-generativeai
notion-client
tenacity
requests
uvloop
httpx
pydantic